Using URI template patterns  page 75
Added a way to display items by category Using URI template patterns.
@RequestMapping("/{category}")

http://localhost:8080/webstore/products/laptop - will now display just
the laptops.

we indicate to the Spring MVC that it is a URI template variable.
According to Spring MVC documentation, a URI template is a URI-like
string that contains one or more variable names. When you substitute
values for these variables, the template becomes a URI.